Remove timetravel stuff from layout reducer.
