+++ client/ducks/layout.js
@@ -28,8 +28,6 @@ export const EMIT_STALL_OCCUPANCIES_UPDATE = 'EMIT_STALL_OCCUPANCIES_UPDATE';
 export const EMIT_FETCH_MARKERS_ERROR = 'EMIT_FETCH_MARKERS_ERROR';
 export const EMIT_CLEAR_CONNECTION_ERRORS = 'EMIT_CLEAR_CONNECTION_ERRORS';
 
-export const EMIT_TIME_TRAVEL_UPDATE = 'EMIT_TIME_TRAVEL_UPDATE';
-
 export const connectSocket = (payload) => ({
   type: CONNECT_SOCKET,
   payload
@@ -54,11 +52,6 @@ export const emitToggleTempScale = (tempScale) => ({
   tempScale
 });
 
-export const emitTimeTravelUpdate = (timeTravelledTo) => ({
-  type: EMIT_TIME_TRAVEL_UPDATE,
-  timeTravelledTo
-});
-
 const initialState = immutable.fromJS({
   meetingRooms: null,
   markers: null,
@@ -66,8 +59,7 @@ const initialState = immutable.fromJS({
   displayLegend: true,
   displayTemp: true, // User UI option for toggling temperature display on map
   enableTemp: false, // Server variable that governs display of temperature UI elements
-  tempScale: FAHRENHEIT,
-  timeTravelledTo: null
+  tempScale: FAHRENHEIT
 });
 
 const layoutReducer = (state = initialState, action) => {
@@ -140,10 +132,6 @@ const layoutReducer = (state = initialState, action) => {
 
     [EMIT_ROOM_MOTION_UPDATE]() {
       return state;
-    },
-
-    [EMIT_TIME_TRAVEL_UPDATE]() {
-      return state.set('timeTravelledTo', action.timeTravelledTo);
     }
   };
