+++ client/middleware/api.js
@@ -1,8 +1,8 @@
 import connectSocket, { send } from './connect-socket';
 import { TIME_TRAVEL_UPDATE } from '../constants';
+import { EMIT_TIME_TRAVEL_UPDATE } from '../ducks/navigation';
 import { CONNECT_SOCKET,
-         EMIT_FETCH_ROOM_STATUSES_ERROR,
-         EMIT_TIME_TRAVEL_UPDATE } from '../ducks/layout';
+         EMIT_FETCH_ROOM_STATUSES_ERROR } from '../ducks/layout';
 
 export default () => (next) => (action) => {
   switch (action.type) {
